refactor: Move `find_closest_match` function from `utils.py` to `custom/utils.py` for better code organization.
