+++ src/backend/base/langflow/custom/custom_component/component.py
@@ -7,7 +7,6 @@ import nanoid  # type: ignore
 import yaml
 from pydantic import BaseModel
 
-from langflow.custom.utils import find_closest_match
 from langflow.graph.state.model import create_state_model
 from langflow.helpers.custom import format_type
 from langflow.schema.artifact import get_artifact_type, post_process_raw
@@ -17,6 +16,7 @@ from langflow.services.tracing.schema import Log
 from langflow.template.field.base import UNDEFINED, Input, Output
 from langflow.template.frontend_node.custom_components import ComponentFrontendNode
 from langflow.utils.async_helpers import run_until_complete
+from langflow.utils.util import find_closest_match
 
 from .custom_component import CustomComponent
