+++ src/backend/base/langflow/custom/utils.py
@@ -567,13 +567,3 @@ def get_instance_name(instance):
     if hasattr(instance, "name") and instance.name:
         name = instance.name
     return name
-
-
-def find_closest_match(string: str, list_of_strings: list[str]) -> str | None:
-    """
-    Find the closest match in a list of strings.
-    """
-    closest_match = difflib.get_close_matches(string, list_of_strings, n=1, cutoff=0.2)
-    if closest_match:
-        return closest_match[0]
-    return None
